+++ Source/DREAM3DLib/IOFilters/YSChoiAbaqusReader.cpp
@@ -209,7 +209,7 @@ void YSChoiAbaqusReader::dataCheck()
     }
   }
 
-  QVector<int> dims(1, 3);
+  QVector<size_t> dims(1, 3);
   m_CellEulerAnglesPtr = cellAttrMat->createNonPrereqArray<DataArray<float>, AbstractFilter, float>(this,  m_CellEulerAnglesArrayName, 0, dims); /* Assigns the shared_ptr<> to an instance variable that is a weak_ptr<> */
   if( NULL != m_CellEulerAnglesPtr.lock().get() ) /* Validate the Weak Pointer wraps a non-NULL pointer to a DataArray<T> object */
   { m_CellEulerAngles = m_CellEulerAnglesPtr.lock()->getPointer(0); } /* Now assign the raw pointer to data from the DataArray<T> object */
