+++ Source/DREAM3DLib/StatisticsFilters/FindFeatureNeighborCAxisMisalignments.cpp
@@ -112,7 +112,7 @@ void FindFeatureNeighborCAxisMisalignments::dataCheck()
   AttributeMatrix* cellFeatureAttrMat = m->getPrereqAttributeMatrix<AbstractFilter>(this, getCellFeatureAttributeMatrixName(), -301);
   if(getErrorCondition() < 0) { return; }
 
-  QVector<int> dims(1, 4);
+  QVector<size_t> dims(1, 4);
   // Feature Data
   m_AvgQuatsPtr = cellFeatureAttrMat->getPrereqArray<DataArray<float>, AbstractFilter>(this, m_AvgQuatsArrayName, -301, dims); /* Assigns the shared_ptr<> to an instance variable that is a weak_ptr<> */
   if( NULL != m_AvgQuatsPtr.lock().get() ) /* Validate the Weak Pointer wraps a non-NULL pointer to a DataArray<T> object */
